refactor!: rename package from karat to coaxer
Renames the Python package, CLI entry, and imports from `karat` to
`coaxer`. The repository is now hosted at github.com/thekevinscott/coaxer.

The old `karat` package remains as a deprecation shim: importing `karat`
or running the `karat` CLI re-exports from `coaxer` and emits a
`DeprecationWarning`. The shim will be removed in a future release.

BREAKING CHANGE: Install `coaxer` (`uv add coaxer`) and switch imports
to `from coaxer import ...`. `from karat import ...` still works but is
deprecated.
