+++ karat/__init__.py
@@ -1,40 +1,21 @@
-"""DSPy language models for Claude Code and OpenAI-compatible endpoints.
+"""Deprecated alias for the :mod:`coaxer` package.
 
-AgentLM routes calls through the Anthropic Agent SDK (Claude Code).
-OpenAILM calls any OpenAI-compatible chat completion API (Ollama, vLLM,
-OpenAI, etc.).
-
-Quick start with Claude Code::
-
-    import dspy
-    from karat import AgentLM
-
-    lm = AgentLM()
-    dspy.configure(lm=lm)
-
-Quick start with Ollama::
-
-    from karat import OpenAILM
-
-    lm = OpenAILM(model="llama3")
-    dspy.configure(lm=lm)
-
-Quick start with OpenAI::
-
-    lm = OpenAILM(model="gpt-4o", base_url="https://api.openai.com/v1", api_key="sk-...")
-
-Load optimized programs saved by the /optimize skill::
-
-    from karat import load_predict
-    classify = load_predict(ClassifyRepo, path="data/optimized.json")
+The project was renamed from ``karat`` to ``coaxer``. Importing
+``karat`` continues to work via re-exports from :mod:`coaxer` but emits
+a :class:`DeprecationWarning`. The shim will be removed in a future
+release; migrate to ``from coaxer import ...``.
 """
 
-from .lm import AgentLM
-from .load_predict import load_predict
-from .openai_lm import OpenAILM
+import warnings
 
-__all__ = [
-    "AgentLM",
-    "OpenAILM",
-    "load_predict",
-]
+warnings.warn(
+    "`karat` has been renamed to `coaxer`. "
+    "Replace `from karat import X` with `from coaxer import X`. "
+    "The `karat` shim will be removed in a future release.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+from coaxer import AgentLM, OpenAILM, load_predict  # noqa: E402
+
+__all__ = ["AgentLM", "OpenAILM", "load_predict"]
